+++ src/main/kotlin/cbnu/io/cbnualramischeduler/business/web/crawling/application/CbnuNotificationCrawler.kt
@@ -1,6 +1,6 @@
 package cbnu.io.cbnualramischeduler.business.web.crawling.application
 
-import cbnu.io.cbnualramischeduler.business.core.domain.notification.Notification
+import cbnu.io.cbnualramischeduler.business.core.domain.notification.entity.Notification
 
 interface CbnuNotificationCrawler {
