+++ src/main/kotlin/cbnu/io/cbnualramischeduler/business/web/scheduler/application/service/NotificationSchedulerService.kt
@@ -2,7 +2,6 @@ package cbnu.io.cbnualramischeduler.business.web.scheduler.application.service
 
 import cbnu.io.cbnualramischeduler.business.web.notification.application.NotificationWriteCommand
 import cbnu.io.cbnualramischeduler.business.web.scheduler.application.NotificationScheduler
-import org.slf4j.LoggerFactory
 import org.springframework.scheduling.annotation.Scheduled
 import org.springframework.stereotype.Component
 
@@ -11,12 +10,14 @@ class NotificationSchedulerService(
     val notificationWriteCommand: NotificationWriteCommand,
 ): NotificationScheduler {
 
-    private val log = LoggerFactory.getLogger(this.javaClass)
+//    private val log = LoggerFactory.getLogger(this.javaClass)
+//    private val log = KotlinLogging.logger {}
+
 
 
     @Scheduled(cron = "0 0/30 * * * ?")
     override fun scheduleNotification() {
-        log.info("=========================스케줄러 시작=========================")
+//        log.info("=========================스케줄러 시작=========================")
         notificationWriteCommand.scheduleSaveTask()
     }
 }
